Add blurhash container component
